Reorganizando mocks e adaptando código necessário
